refactor(redirect): move GA logging types to redirect
