+++ src/server/modules/redirect/RedirectController.ts
@@ -6,8 +6,8 @@ import parseDomain from '../../util/domain'
 import { DependencyIds, ERROR_404_PATH } from '../../constants'
 import { AnalyticsLoggerService, RedirectService } from './services'
 import { RedirectType } from '.'
-import { EventAction, EventCategory } from '../../services/analytics/types/enum'
-import { createPageViewHit } from '../../services/analytics'
+import { EventAction, EventCategory } from './ga/types/enum'
+import { createPageViewHit } from './ga'
 
 const TRANSITION_PATH = 'transition-page.ejs'
 const GTAG_PATH = 'redirect.ejs'
