+++ src/server/modules/redirect/ga/index.ts
@@ -1,7 +1,7 @@
 import express from 'express'
 import { v4 as uuidv4 } from 'uuid'
-import { gaTrackingId, ogUrl } from '../../config'
-import getIp from '../../util/request'
+import { gaTrackingId, ogUrl } from '../../../config'
+import getIp from '../../../util/request'
 import IGaPageViewForm from './types/IGaPageViewForm'
 import IGaCoreForm from './types/IGaCoreForm'
 import { GaHitVariant } from './types/enum'
